NeuroLang/pyproject.toml
[project]
name = "neurolang"
version = "0.1.0"
description = "NeuroLang - kompilator autorskiego języka programowania do skryptów PyTorch"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Kacper Lipiec" },
    { name = "Przemysław Kondrat" },
]
dependencies = [
    "lark>=1.1.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchmetrics>=0.11.0",
    "torchview>=0.2.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
neurolang = "src.cli.compile:main"
neurolang-ast = "src.cli.show_ast:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
